api/requirements.txt
fastapi
pillow
torch
torchvision
torchaudio
python-multipart
soundfile
pandas
annoy
numpy
librosa
matplotlib